backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
Pillow>=10.1.0
python-decouple>=3.8
django-filter>=23.3
django-jet>=1.3.4
djangorestframework-simplejwt>=5.3.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0